requirements.txt
# ============================================================================
# AxiomOS Labs 项目依赖
# ----------------------------------------------------------------------------

# 核心运行时依赖（运行系统必需）
openai>=1.12.0                # OpenAI 兼容 API 客户端（用于 DeepSeek 等 LLM）
python-dotenv>=1.0.0         # 环境变量管理（从 .env 文件加载配置）
mcp>=0.1.0                   # Model Context Protocol 客户端/服务器库（远程技能调用）

# 开发与测试依赖（用于本地开发、测试、代码质量）
pytest>=7.4.0                # 测试框架
pytest-asyncio>=0.21.0       # 异步测试支持
pytest-cov>=4.1.0            # 测试覆盖率报告
black>=23.0.0                # 代码格式化（强制执行统一风格）
isort>=5.12.0                # 导入语句排序
mypy>=1.5.0                  # 静态类型检查
flake8>=6.0.0                # 代码风格与错误检查
types-python-dotenv>=1.0.0   # python-dotenv 的类型提示

# 可选依赖（增强功能，非必需）
pydantic>=2.5.0              # 数据验证与设置管理（未来配置升级）
requests>=2.31.0             # 通用 HTTP 客户端（备用）
aiohttp>=3.9.0               # 异步 HTTP 客户端（高性能场景）
colorlog>=6.7.0              # 彩色日志输出（提升可读性）
colorama>=0.4.6              # 跨平台彩色终端输出（兼容 Windows）
typing-extensions>=4.8.0     # 新版类型提示支持（Python 3.8 兼容）

# 系统级依赖（非 Python 包）
# Fast‑Downward 规划器作为 git 子模块提供，需要以下系统工具：
#   - g++ (>= 9.0) 或 clang++
#   - cmake (>= 3.10)
#   - flex
#   - bison
#   - python3-dev
# 请参阅 downward/README.md 获取完整编译指南。

# ----------------------------------------------------------------------------
# 安装说明
# ----------------------------------------------------------------------------
# 仅安装核心依赖：
#   pip install openai python-dotenv mcp
#
# 安装核心 + 可选依赖：
#   pip install -r requirements.txt --upgrade
#
# 安装开发环境（包含测试与格式化工具）：
#   pip install pytest black isort mypy flake8 types-python-dotenv
#
# 生产部署建议仅安装核心依赖。
# ============================================================================